Time: 1 ms (99.73%) | Memory: 45.8 MB (66.75%) - LeetSync

+++ 260-single-number-iii/single-number-iii.java
@@ -1,11 +1,11 @@
 class Solution {
     public int[] singleNumber(int[] nums) {
         
-        int  exor=0;
+        long  exor=0;
         for(int i=0;i<nums.length;i++){
             exor^=nums[i];
         }
-        int rightmostbit=(exor&exor-1)^exor;
+        long rightmostbit=(exor&exor-1)^exor;
         int b1=0,b2=0;
         for(int i=0;i<nums.length;i++){
             if((rightmostbit & nums[i])==0){
